+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_params.py
@@ -15,7 +15,8 @@ from azure.mgmt.containerregistry.v2018_09_01.models import (
     PolicyStatus,
     RunStatus,
     TaskStatus,
-    BaseImageTriggerType
+    BaseImageTriggerType,
+    DefaultAction
 )
 from azure.mgmt.containerregistry.v2018_02_01_preview.models import (
     BuildTaskStatus,
@@ -74,6 +75,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('location', arg_type=get_location_type(self.cli_ctx))
         c.argument('tags', arg_type=tags_type)
         c.argument('registry_name', options_list=['--name', '-n'], help='The name of the container registry. You can configure the default registry name using `az configure --defaults acr=<registry name>`', completer=get_resource_name_completion_list(REGISTRY_RESOURCE_TYPE), configured_default='acr')
+        c.argument('tenant_suffix', options_list=['--suffix'], help="The tenant suffix in registry login server. You may specify '--suffix tenant' if your registry login server is in the format 'registry-tenant.azurecr.io'. Applicable if you\'re accessing the registry from a different subscription or you have permission to access images but not the permisson to manage the registry resource.")
         c.argument('storage_account_name', help='Provide the name of an existing storage account if you\'re recreating a container registry over a previous registry created storage account. Only applicable to Classic SKU.', completer=get_resource_name_completion_list(STORAGE_RESOURCE_TYPE))
         c.argument('sku', help='The SKU of the container registry', arg_type=get_enum_type(MANAGED_REGISTRY_SKU + CLASSIC_REGISTRY_SKU))
         c.argument('admin_enabled', help='Indicates whether the admin user is enabled', arg_type=get_three_state_flag())
@@ -88,6 +90,11 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
         c.argument('no_wait', help="Do not wait for the run to complete and return immediately after queuing the run.", action='store_true')
         c.argument('no_format', help="Indicates whether the logs should be displayed in raw format", action='store_true')
         c.argument('os_type', options_list=['--os'], help='The operating system type required for the build.', arg_type=get_enum_type(OsType))
+
+    for scope in ['acr create', 'acr update']:
+        with self.argument_context(scope, arg_group='Network Rule') as c:
+            c.argument('default_action', arg_type=get_enum_type(DefaultAction),
+                       help='Default action to apply when no rule matches. Only applicable to Premium SKU.')
 
     with self.argument_context('acr import') as c:
         c.argument('source', help="The source identifier in the format '[registry.azurecr.io/]repository[:tag]' or '[registry.azurecr.io/]repository@digest'.")
@@ -251,3 +258,8 @@ def load_arguments(self, _):  # pylint: disable=too-many-statements
     with self.argument_context('acr helm push') as c:
         c.positional('chart_package', help="The helm chart package.", completer=FilesCompleter())
         c.argument('force', help='Overwrite the existing chart package.', action='store_true')
+
+    with self.argument_context('acr network-rule') as c:
+        c.argument('subnet', help='Name or ID of subnet. If name is supplied, `--vnet-name` must be supplied.')
+        c.argument('vnet_name', help='Name of a virtual network.')
+        c.argument('ip_address', help='IPv4 address or CIDR range.')