+++ src/command_modules/azure-cli-eventgrid/azure/cli/command_modules/eventgrid/commands.py
@@ -21,12 +21,12 @@ def load_command_table(self, _):
     )
 
     with self.command_group('eventgrid topic', topics_mgmt_util, client_factory=topics_factory) as g:
-        g.command('create', 'create_or_update')
         g.show_command('show', 'get')
         g.command('key list', 'list_shared_access_keys')
         g.command('key regenerate', 'regenerate_key')
         g.command('delete', 'delete')
         g.custom_command('list', 'cli_topic_list')
+        g.custom_command('create', 'cli_topic_create_or_update')
         g.generic_update_command('update',
                                  getter_name='get',
                                  setter_name='update',