+++ src/command_modules/azure-cli-hdinsight/azure/cli/command_modules/hdinsight/_client_factory.py
@@ -17,6 +17,13 @@ def cf_storage(cli_ctx, *_, **__):
     return get_mgmt_service_client(cli_ctx, StorageManagementClient)
 
 
+def cf_network(cli_ctx, aux_subscriptions=None, **_):
+    from azure.cli.core.profiles import ResourceType
+    from azure.cli.core.commands.client_factory import get_mgmt_service_client
+    return get_mgmt_service_client(cli_ctx, ResourceType.MGMT_NETWORK,
+                                   aux_subscriptions=aux_subscriptions)
+
+
 def cf_hdinsight(cli_ctx, *_, **__):
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
     from azure.mgmt.hdinsight import HDInsightManagementClient