+++ src/azure-cli-core/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+2.0.56
+++++++
+* auth: enable tenant level account for managed service identity
 
 2.0.55
 ++++++